refactor: modularize similarity matrixes calc
